Fix a bug where permutation was shuffling spaces
The last version had a logical error where spaces were being
treated as part of the plaintext and accordingly being shuffled.

Now the algorithm skips these spaces, while still preserving their
location in the output.

Also, the README is now a .tex file that is converted to a PDF (for
my class assignment) using pdflatex and Github Flavored Markdown for
the public copy of this repository.