Update permutation implementation for clarity
In fixing the permutation encryption so that it would ignore the
spaces, the function definition looked horribly convoluted.  I had
a thought in class the other day, that the spaces could be removed,
then the encryption/decryption block-by-block, then add the spaces
back.

All of that instead of trying to do both of these steps together.
Both versions still ended up looping through the full plaintext
twice, but the new version uses two separate loops.